+++ Android/GameProtection/protectlib/src/main/java/com/antiaddiction/sdk/AntiAddictionCore.java
@@ -36,6 +36,9 @@ public class AntiAddictionCore {
     private static boolean hasLogin = false;
     private static boolean isForeign = false;
 
+    private static final String ANTI_ADDICTION_UN_REAL_NAME_LIMIT_TITLE = "健康游戏提示";
+    private static final String ANTI_ADDICTION_UN_REAL_NAME_LIMIT_CONTENT = "您的账号未完成实名认证，为了符合国家相关规定，不影响您的游戏体验，请尽快完善实名信息。";
+
     private static Handler mainHandler = new Handler(Looper.getMainLooper()) {
         @Override
         public void handleMessage(Message msg) {
@@ -93,6 +96,7 @@ public class AntiAddictionCore {
                     if (null != protectCallBack) {
                         protectCallBack.onAntiAddictionResult(AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME, type);
                     }
+                    currentUser = null;
                     break;
                 case AntiAddictionKit.CALLBACK_CODE_CHAT_LIMIT:
                     if (null != protectCallBack) {
@@ -190,6 +194,7 @@ public class AntiAddictionCore {
             //登录
             if (currentUser == null) {
                 if(!AntiAddictionKit.getFunctionConfig().getSupportSubmitToServer()) {
+                    /* ------------单机版-------------- */
                     currentUser = UserDao.getUser(activity, userId);
                     //本地未存储用户信息
                     if (null == currentUser) {
@@ -204,6 +209,7 @@ public class AntiAddictionCore {
                     LogUtil.logd("getUser info = " + getCurrentUser().toJsonString());
                     checkUser();
                 }else{
+                    /* ------------联网版-------------- */
                     UserService.getUserInfo(userId, userType,new Callback() {
                         @Override
                         public void onSuccess(JSONObject response) {
@@ -517,7 +523,7 @@ public class AntiAddictionCore {
     }
 
     static String getSdkVersion() {
-        return "1.1.2";
+        return "1.1.3";
     }
 
     /**
@@ -539,109 +545,141 @@ public class AntiAddictionCore {
                 boolean needShowRealName = AntiAddictionKit.getFunctionConfig().getUseSdkRealName();
                 if (null != strictData) {
                     try {
-                        final int restrictType = strictData.getInt("restrictType");//0无限制 1 宵禁 2 时长限制
+                        final int restrictType = strictData.getInt("restrictType");//(old: 0无限制 1 宵禁 2 时长限制) new rules---> 0: 不受限制; 1:已实名不在可玩时间段限制; 2: 已实名在可玩时间内时长限制 ; 3: 未实名限制
                         final String title = strictData.optString("title");
                         final String content = strictData.optString("description");
                         final int remainTime = strictData.getInt("remainTime");
-                        if (restrictType == 0) {
-                            if(getCurrentUser().getAccountType() != AntiAddictionKit.USER_TYPE_UNKNOWN) {
-                                getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
-                            }else{
-                                String tip = PlayLogService.generateGuestLoginTip(remainTime);
-                                AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_NO_LIMIT, title,
-                                        tip, 2, new OnResultListener() {
-                                            @Override
-                                            public void onResult(int type, String msg) {
-                                                if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
-                                                    logout();
-                                                } else {
-                                                    if (type != 0) {
-                                                        getCallBack().onResult(type, msg);
-                                                    }
-                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
-                                                }
-                                            }
-                                        }, needShowRealName);
 
-                            }
+                        int limitState = -1; // 防沉迷弹窗出现的限制类型: -1 不受限制, 1.提示未成年人当前不在可玩时间段，2.提示未成年人当前剩余可玩时间, 3. 提示需要实名
+                        if (restrictType == 3) {
+                            limitState = AccountLimitTip.STATE_UN_REAL_NAME_LIMIT;
+                        } else if (restrictType == 2) {
+                            limitState = AccountLimitTip.STATE_MINOR_TIME_MAX_LIMIT;
+                        } else if (restrictType == 1) {
+                            limitState = AccountLimitTip.STATE_MINOR_TIME_RANGE_LIMIT;
+                        }
+
+                        if (limitState == -1) {
+                            // 不受限制且已实名的, 可成功
+                            getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
                         } else {
-                            if (restrictType == 1) {
-                                if (remainTime <= 0) {
-                                    AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_STRICT,
-                                            title, content, 1, new OnResultListener() {
-                                                @Override
-                                                public void onResult(int type, String msg) {
-                                                    if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
-                                                        logout();
-                                                    }
-                                                }
-                                            });
-                                } else {
-                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
-                                }
-                            } else if (restrictType == 2) { //游戏时间限制或提醒 未成年人
-                                if (getCurrentUser().getAccountType() > AntiAddictionKit.USER_TYPE_UNKNOWN) {
-                                    if (remainTime <= 0) { //无游戏时间额度
-                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_STRICT,
-                                                title, content, 1, new OnResultListener() {
-                                                    @Override
-                                                    public void onResult(int type, String msg) {
-                                                        if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
-                                                            logout();
-                                                        }
-                                                    }
-                                                });
+                            AccountLimitTip.showAccountLimitTip(limitState, title, content, 3, new OnResultListener() {
+                                @Override
+                                public void onResult(int type, String msg) {
+                                    if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
+                                        logout();
+                                    } else if (type == AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME) {
+                                        getCallBack().onResult(type, msg);
                                     } else {
-//                                AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_NO_LIMIT,
-//                                        title, content, 1);
+                                        // 单机版防沉迷只会实名成功不会实名失败
+                                        if (type != 0) {
+                                            getCallBack().onResult(type, msg);
+                                        }
                                         getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
                                     }
-                                } else { //游客
-                                    if (remainTime <= 0) { //无游戏时间额度
-                                        OnResultListener onResultListener = new OnResultListener() {
-                                            @Override
-                                            public void onResult(int type, String msg) {
-                                                if (type == AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME) {
-                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME, msg);
-                                                    return;
-                                                }
-                                                if (type != AntiAddictionKit.CALLBACK_CODE_REAL_NAME_SUCCESS) {
-                                                    if (type == AntiAddictionKit.CALLBACK_CODE_REAL_NAME_FAIL) {
-                                                        // getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_REAL_NAME_FAIL,"");
-                                                    }
-                                                    logout();
-                                                } else {
-                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
-                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_USER_TYPE_CHANGED, "");
-                                                }
-                                            }
-                                        };
-                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_LIMIT, title,
-                                                content, 1, onResultListener, needShowRealName);
-                                    } else {
-                                        String tip = PlayLogService.generateGuestLoginTip(remainTime);
-                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_NO_LIMIT, title,
-                                                tip, 2, new OnResultListener() {
-                                                    @Override
-                                                    public void onResult(int type, String msg) {
-                                                        if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
-                                                            logout();
-                                                        } else {
-                                                            if (type != 0) {
-                                                                if(type == AntiAddictionKit.CALLBACK_CODE_REAL_NAME_SUCCESS){
-                                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_USER_TYPE_CHANGED,"");
-                                                                }else {
-                                                                    getCallBack().onResult(type, msg);
-                                                                }
-                                                            }
-                                                            getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
-                                                        }
-                                                    }
-                                                }, needShowRealName);
-                                    }
                                 }
-                            }
+                            }, needShowRealName);
                         }
+
+//                        if (restrictType == 0) {
+//                            if(getCurrentUser().getAccountType() != AntiAddictionKit.USER_TYPE_UNKNOWN) {
+//                                getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                            }else{
+//                                String tip = PlayLogService.generateGuestLoginTip(remainTime);
+//                                AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_NO_LIMIT, title,
+//                                        tip, 2, new OnResultListener() {
+//                                            @Override
+//                                            public void onResult(int type, String msg) {
+//                                                if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
+//                                                    logout();
+//                                                } else {
+//                                                    if (type != 0) {
+//                                                        getCallBack().onResult(type, msg);
+//                                                    }
+//                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                                                }
+//                                            }
+//                                        }, needShowRealName);
+//
+//                            }
+//                        } else {
+//                            if (restrictType == 1) {
+//                                if (remainTime <= 0) {
+//                                    AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_STRICT,
+//                                            title, content, 1, new OnResultListener() {
+//                                                @Override
+//                                                public void onResult(int type, String msg) {
+//                                                    if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
+//                                                        logout();
+//                                                    }
+//                                                }
+//                                            });
+//                                } else {
+//                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                                }
+//                            } else if (restrictType == 2) { //游戏时间限制或提醒 未成年人
+//                                if (getCurrentUser().getAccountType() > AntiAddictionKit.USER_TYPE_UNKNOWN) {
+//                                    if (remainTime <= 0) { //无游戏时间额度
+//                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_STRICT,
+//                                                title, content, 1, new OnResultListener() {
+//                                                    @Override
+//                                                    public void onResult(int type, String msg) {
+//                                                        if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
+//                                                            logout();
+//                                                        }
+//                                                    }
+//                                                });
+//                                    } else {
+////                                AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_CHILD_ENTER_NO_LIMIT,
+////                                        title, content, 1);
+//                                        getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                                    }
+//                                } else { //游客
+//                                    if (remainTime <= 0) { //无游戏时间额度
+//                                        OnResultListener onResultListener = new OnResultListener() {
+//                                            @Override
+//                                            public void onResult(int type, String msg) {
+//                                                if (type == AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME) {
+//                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_OPEN_REAL_NAME, msg);
+//                                                    return;
+//                                                }
+//                                                if (type != AntiAddictionKit.CALLBACK_CODE_REAL_NAME_SUCCESS) {
+//                                                    if (type == AntiAddictionKit.CALLBACK_CODE_REAL_NAME_FAIL) {
+//                                                        // getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_REAL_NAME_FAIL,"");
+//                                                    }
+//                                                    logout();
+//                                                } else {
+//                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_USER_TYPE_CHANGED, "");
+//                                                }
+//                                            }
+//                                        };
+//                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_LIMIT, title,
+//                                                content, 1, onResultListener, needShowRealName);
+//                                    } else {
+//                                        String tip = PlayLogService.generateGuestLoginTip(remainTime);
+//                                        AccountLimitTip.showAccountLimitTip(AccountLimitTip.STATE_ENTER_NO_LIMIT, title,
+//                                                tip, 2, new OnResultListener() {
+//                                                    @Override
+//                                                    public void onResult(int type, String msg) {
+//                                                        if (type == AntiAddictionKit.CALLBACK_CODE_SWITCH_ACCOUNT) {
+//                                                            logout();
+//                                                        } else {
+//                                                            if (type != 0) {
+//                                                                if(type == AntiAddictionKit.CALLBACK_CODE_REAL_NAME_SUCCESS){
+//                                                                    getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_USER_TYPE_CHANGED,"");
+//                                                                }else {
+//                                                                    getCallBack().onResult(type, msg);
+//                                                                }
+//                                                            }
+//                                                            getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
+//                                                        }
+//                                                    }
+//                                                }, needShowRealName);
+//                                    }
+//                                }
+//                            }
+//                        }
                     } catch (Exception e) {
                         getCallBack().onResult(AntiAddictionKit.CALLBACK_CODE_LOGIN_SUCCESS, "");
                     }
@@ -655,6 +693,14 @@ public class AntiAddictionCore {
             }
         });
 
+    }
+
+    private static boolean isUnRealNameLimit(boolean needShowRealName) {
+        if (currentUser.getAccountType() == AntiAddictionKit.USER_TYPE_UNKNOWN) {
+
+            return true;
+        }
+        return false;
     }
 
     /**